chore: refactor authentication failure redirect and retry
