+++ app/server/appsmith-server/src/main/java/com/appsmith/server/authentication/helpers/AuthenticationFailureRetryHandler.java
@@ -1,0 +1,3 @@
+package com.appsmith.server.authentication.helpers;
+
+public interface AuthenticationFailureRetryHandler extends AuthenticationFailureRetryHandlerCE {}
