+++ app/server/appsmith-server/src/main/java/com/appsmith/server/authentication/helpers/AuthenticationFailureRetryHandlerCE.java
@@ -1,0 +1,10 @@
+package com.appsmith.server.authentication.helpers;
+
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.server.WebFilterExchange;
+import reactor.core.publisher.Mono;
+
+public interface AuthenticationFailureRetryHandlerCE {
+    Mono<Void> retryAndRedirectOnAuthenticationFailure(
+            WebFilterExchange webFilterExchange, AuthenticationException exception);
+}
