+++ app/server/appsmith-server/src/main/java/com/appsmith/server/authentication/helpers/AuthenticationFailureRetryHandlerImpl.java
@@ -1,0 +1,7 @@
+package com.appsmith.server.authentication.helpers;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthenticationFailureRetryHandlerImpl extends AuthenticationFailureRetryHandlerCEImpl
+        implements AuthenticationFailureRetryHandler {}
